Add files via upload
This project focuses on analyzing the sales performance of various products in an online retail store, with a specific emphasis on identifying the top-selling items. Using exploratory data analysis (EDA), the project examines the quantity of products sold to uncover patterns, trends, and customer preferences. The goal is to provide insights that can support decision-making in product development, marketing strategies, and inventory management. The dataset includes a list of products and their corresponding sales volumes, helping to highlight the most popular items among customers.